text_into_commands/pyproject.toml
[tool.poetry]
name = "text-into-commands"
version = "0.1.0"
description = ""
authors = ["Irlen Turlykhanov <irlen.turlykhanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
langchain-anthropic = "^0.3.10"
langchain = "^0.3.20"
grpcio = "^1.71.0"
langchain-openai = "^0.3.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
